+++ src/contentTypeConfig.js
@@ -218,11 +218,7 @@ export const showsLessonTypes = [
   'spotlight',
   'performance',
 ]
-export const entertainmentLessonTypes = [
-  'special',
-  'documentary-lesson',
-  ...showsLessonTypes,
-]
+export const entertainmentLessonTypes = ['special', 'documentary-lesson', ...showsLessonTypes]
 export const collectionLessonTypes = [...coursesLessonTypes, ...showsLessonTypes]
 
 export const lessonTypesMapping = {
@@ -401,6 +397,9 @@ export let contentTypeConfig = {
   },
   'guided-course': {
     includeChildFields: true,
+  },
+  'course-collection': {
+    individualLessonsTypes: true,
   },
   course: {
     fields: [